pyproject.toml
[tool.uv]
dev-dependencies = []

[project]
name = "mcp-retrieval-engine"
version = "0.1.0"
description = "A production-ready Model Context Protocol (MCP) server implementing a local RAG pipeline."
requires-python = ">=3.10"
dependencies = [
  "fastmcp",
  "langchain",
  "langchain-community",
  "langchain-huggingface",
  "chromadb",
  "sentence-transformers",
  "pypdf",
]